Integrate flag printing instead of using PrintDefaults.
When rendering flags for help print long flag names prefixed with "--" instead
of "-" to fit the expectations of users from other languages.

Also allow flags to be marked PRIVATE: by a prefix on their usage string, to
exclude them from help listings. If all the flags on a flag set have this mark,
no flag help is displayed at all for the command. Private-marked flags are
still parsed normally.