Improve the output of the "version" command.
Make more effort to discover version information, even if vcs.* tags are not
available in the build info because the program was installed "remotely", e.g.,
via "go install" without a repository present.

Also add a --json flag to the version command, allowing these data to be
written out in machine-readable form.